Added cast for compatibility with newer compilers

+++ Algebras/IMEX/ImExAlgebra.cpp
@@ -9054,7 +9054,7 @@ int splitShpVMT(Word* args, Word& result,
       double maxX=0;
       double maxY=0;
       while(   in.good() && (recno < (uint32_t) maxnorec) 
-            && in.tellg()<fileSize){
+            && (size_t)in.tellg()<fileSize){
          in.seekg(4,ios::cur); // ignore original record number
          uint32_t cl = readBigInt32(in);
          char* record = new char[2*cl];
